Created a grid to show selected letters. Created buttons to insert letters into the grid.
Added a 6 by 5 grid of gridSquares, that display a letter upon a button being clicked by the player. Also added 3 rows of kbSquares, representing the buttons that can be clicked on a keyboard. When a kbSquare button is clicked, the letter on that button is displayed on the next available gridSquare. DEL and RET buttons do not yet have functionality.