Gave functionality to DEL and RET buttons
Made it so that the DEL button will set the last changed gridSquare to null. This will only work up until the beginning of the current line. Made it so that the RET key is only a valid button press if the last changed gridSquare is at the end of a line. Additionally, users are forced to either click the DEL or RET key when at the end of a line.