Completed RET functionality. Squares change colors and added game end message
The RET button is fully functional now. When the RET button is clicked, the color of the squares is updated. If the player wins the game, a win message is displayed with the number of guesses required. If the player loses, a losing message is shown along with what the word was. Once the game is over, all button functionality is disabled.