Created a set of possible words, rather than hardcoding the targetWord
Created a set of random words that can be the targetWord for players to guess. This target word is randomly chosen once at the beginning of the game. Because we now have a set of words, we can enforce the restriction that players must enter a valid word contained in the set, rather than any sequence of 5 letters. If the user enters an invalid word, a message is displayed to enter a new word. In the future, this set will be replaced with a modified version of linux.words, which is called wordleWords.txt.