Colored keyboard according to user guesses
Made it so that the keyboard squares are colored as the game progresses. Similar to the game board squares, the keyboard squares are colored grey if the corresponding letter does not appear in the target word. Also, the square is colored green on the keyboard if the user correctly guessed the letter's position in the word. In order for the square to be colored yellow, more logic was required, due to the possibility of both user guesses and target words containing duplicate letters. The behavior we chose to implement is as follows: If the user guess contains a duplicate letter, but the target word only contains one instance of that letter AND both guesses are in the incorrect position, the keyboard is colored yellow (despite the fact that one square will be yellow and one will be grey on the gameboard). Similarly, if the target word contains two instances of a letter, and the user has correctly guessed the position of one but incorrecly guessed the position of a second, the keyboard square is displayed as green (despite one square being green and one square being yellow on the game board).