Refactored code to support Dynamic PlayGame
Refactored checkStatus to use Maps in a parallel manner to the Python Implementation, storing the letters of targetWord in a Map so as to check for yellows

Refactored generatedWord to accomodate letterRestrictions, specificRequirements, and varying word lengths for future implementations. letterRestrictions and specificRequirements will be lists containing letters not to be included, and letters that must or must not (in the case of ~letter) go in specific spots in the word. Then, a random word is chosen

Refactored playGame to make dyanmic. Current implementation doesn't allow the user to actually input these values, bu the code supports varying word length, number of guesses, letterRestrictions, and specficRequirements. Changed the definitions of components like gridSquares to both simplify the code and to support the new gridRows Map, which stores the squares to be rendered for each guess, allowing dynamic input. gridSquares and newGridSquares now store 'boardsquare' jsx elements, with extractChildren written to extract the actual letters from these. Changed the function rendering the gridSquares to accomodate gridRows

Note: Currently, gridRows has a corresponding React component gridRowsCopy, because making gridRows a component caused a 'too many re-renders' error. Also, colorArr and numGuesses weren't being updated when their set functions were called, so these are manually updated in playGame. This may cause issues later, but seems to work for now