Implemented customGame button
Implemented a new button on the homescreen to allow the user to play a custom game. Added new components showHomeScreen, customGameInitializer, etc in order to separate the display of the home screen, custom game initialization screen, and the actual game. Currently, the html is very bare bones and has no checks for improper input. Next step is to add labels, restrictions to user input, and change the form of the 'specific requirements' button