Intermediate update in adding restrictions/labels for custom games
In this update, I added labels to each parameter, as well as a new way of inputting specificRestrictions, where the user will put the restriction for each letter in it's respective box. This is not fully functional yet, however, as there needs to be some way of chaning the number of boxes dynamically to match word Length.