Added more possible target words
Added more words that could be chosen as the target words. This also added more words that the user could guess. This was done by adding a seperate .jsx file, which holds a set of these words. This list of words is all of the 5 letter words from linux.words, with duplicates removed.