Optimized generateWord, changed max on sliders to 15
Optimized generate word by removing work from the nested loop, implementing checkList to check which indices of SpecificRequirements are meaningful, and directly checking those indices. Also, if LetterRestrictions are empty, they're skipped. This seems to speedup the code substantially for long wordLength, but maybe mild slowdown for shorter words. Overall, it seems to be an improvement

Changed max on sliders to 15, reflecting that our API for generateWordList only goes to 15-long words